Remove workaround
SameSite is a reasonably robust defense against some classes of cross-site request forgery (CSRF) attacks. That's why we cannot recommend to switch it to None even as temporary workaround. 

Temporary workaround must be removed and merchants must be directed to follow the best practice and appropriate solution, which is contacting third party service provided to fix issue accordingly.